Basic implementation for Presenter.
- It is incomplete because we still need some injectors.
- I'm also not sure about the placement of some stuff in terms of abstraction levels, so we will eventually have to check this out as a group.
