+++ app/src/main/java/ayds/newyork/songinfo/moredetails/data/DataBase.kt
@@ -5,7 +5,7 @@ import android.content.Context
 import android.database.Cursor
 import android.database.sqlite.SQLiteDatabase
 import android.database.sqlite.SQLiteOpenHelper
-import ayds.newyork.songinfo.moredetails.domain.ArtistInfo
+import ayds.newyork.songinfo.moredetails.domain.entities.ArtistInfo
 
 class DataBase(context: Context): SQLiteOpenHelper(context, SQLITE_OPEN_HELPER_NAME, null, 1) {
     companion object {
