+++ app/src/main/java/ayds/newyork/songinfo/moredetails/domain/entities/ArtistInfo.kt
@@ -1,7 +1,7 @@
 package ayds.newyork.songinfo.moredetails.domain.entities
 
 data class ArtistInfo(
-    var url: String? = null,
-    var info: String? = null,
+    var url: String? = "",
+    var info: String? = "",
     var isLocallyStored: Boolean = false,
 )
