+++ app/src/main/java/ayds/newyork/songinfo/moredetails/domain/repository/ArtistInfoRepository.kt
@@ -3,5 +3,5 @@ package ayds.newyork.songinfo.moredetails.domain.repository
 import ayds.newyork.songinfo.moredetails.domain.entities.ArtistInfo
 
 interface ArtistInfoRepository {
-    fun getArtistInfo(artistName: String): ArtistInfo?
+    fun getArtistInfo(artistName: String): ArtistInfo
 }
