+++ app/src/main/java/ayds/newyork/songinfo/moredetails/presentation/MoreDetailsUiEvent.kt
@@ -1,0 +1,5 @@
+package ayds.newyork.songinfo.moredetails.presentation
+
+sealed class MoreDetailsUiEvent {
+    object GetArtistInfo: MoreDetailsUiEvent()
+}
